pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "numpy>=2.3.2",
    "openai>=1.101.0",
    "pdfminer-six>=20250506",
    "pydantic-settings>=2.10.1",
    "pypdf>=6.0.0",
    "python-multipart>=0.0.20",
    "structlog>=25.4.0",
    "tiktoken>=0.11.0",
    "uvicorn[standard]>=0.35.0",
    "websockets>=15.0.1",
    "wsproto>=1.2.0",
]
